substitute real(8) by real64
There still was use of `real(8)`.  Double precision `real64` by
`iso_fortran_env` however is independent of the compiler engaged.
Today, it isn't too expensive either to monitor the time of
computation.

File I/O, validity check, provision of output of a grid read from
file (check with diff -s):  results did not change.

+++ app/main.f90
@@ -16,16 +16,17 @@
 ! If not, see <http://www.gnu.org/licenses/>.
 !------------------------------------------------------------------------------
 ! Contributed by Vincent Magnin, 2006-11-27
-! Last modifications: 2023-07-15
+! Last modifications: 2023-07-21
 !------------------------------------------------------------------------------
 
 program main
+    use iso_fortran_env, only: dp => real64
     use sudoku
 
     implicit none
     ! Variables locales :
     integer(1), dimension(1:9, 1:9) :: grille    ! (ligne,colonne)
-    real(8)       :: Debut,Fin    ! Pour mesurer la durée de calcul
+    real(kind=dp)    :: Debut,Fin    ! Pour mesurer la durée de calcul
     integer(1)    :: choix
     integer(1)    :: nvides    ! Nombre de cases à vider
     character(50) :: fichier    ! Nom du fichier .txt
